Phase 2 WIP: Migrate Temperature types (Celsius/Kelvin) to f64 (partial)

+++ crates/core/src/core_types/atmospheric.rs
@@ -37,8 +37,8 @@ pub(crate) fn simulate_plume_rise(grid: &mut SimulationGrid, source_positions: &
 
             if source_temp > *grid.ambient_temperature + 50.0 {
                 // Calculate plume rise velocity
-                let temp_excess = source_temp - *grid.ambient_temperature;
-                let buoyancy_vel = (2.0 * 9.81 * temp_excess / *grid.ambient_temperature).sqrt();
+                let temp_excess = f32::from(source_temp - *grid.ambient_temperature);
+                let buoyancy_vel = (2.0 * 9.81 * temp_excess / f32::from(*grid.ambient_temperature)).sqrt();
 
                 // Transport to cells above
                 let rise_distance = buoyancy_vel * dt;
